requirements.txt
selenium
webdriver-manager
python-dotenv
supabase
requests
beautifulsoup4
